# api/openapi.yaml
openapi: 3.1.0
info:
  title: AtApp Gateway API
  description: API Gateway interface for client-service interactions in AtApp college attendance system.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post:
      summary: Authenticate user
      description: Logs in a student, teacher, or admin and returns JWT claims containing user role and college_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: teacher@mit.edu
                password:
                  type: string
                  format: password
                  example: securepassword123
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken:
                    type: string
                    example: d310702d-6be6-4444-8848-18e404bf7c50
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                        example: Dr. Jane Doe
                      email:
                        type: string
                        example: teacher@mit.edu
                      role:
                        type: string
                        enum: [student, teacher, admin]
                        example: teacher
                      collegeId:
                        type: string
                        format: uuid
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Access token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid or expired refresh token
  /auth/password-reset/request:
    post:
      summary: Request a password reset
      description: >
        Always returns 200 with a generic message whether or not the email is
        registered, to avoid leaking which accounts exist. The reset token is
        currently logged server-side rather than emailed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Request accepted (does not confirm the account exists)
        '429':
          description: Rate limited
  /auth/password-reset/confirm:
    post:
      summary: Confirm a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Invalid or expired reset token, or password too short
        '429':
          description: Rate limited
...
  /teacher/attendance/overrides:
    post:
      summary: Manually override student attendance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
                - studentId
                - status
                - reason
              properties:
                sessionId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [overridden_present, overridden_absent]
                reason:
                  type: string
                  example: "Student was present but forgot phone"
      responses:
        '200':
          description: Override successfully recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordId:
                    type: string
                    format: uuid
                  status:
                    type: string
        '400':
          description: Invalid status value
        '404':
          description: Session not found or not taught by this teacher
  /teacher/sections:
    get:
      summary: List sections taught by the authenticated teacher
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of teacher's sections
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    sectionId:
                      type: string
                      format: uuid
                    term:
                      type: string
                    courseName:
                      type: string
                    courseCode:
                      type: string
                    classroomBssid:
                      type: string
                      nullable: true
                    classroomGeofenceLat:
                      type: number
                      nullable: true
                    classroomGeofenceLng:
                      type: number
                      nullable: true
                    classroomGeofenceRadiusM:
                      type: number
                      nullable: true
  /teacher/sessions/{sessionId}/roster:
    get:
      summary: Get the attendance roster for a class session
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Roster with per-student check-in status
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    studentId:
                      type: string
                      format: uuid
                    name:
                      type: string
                    email:
                      type: string
                    status:
                      type: string
                      enum: [present, absent, overridden_present, overridden_absent, pending]
                    recordedAt:
                      type: string
                      format: date-time
                      nullable: true
        '403':
          description: This teacher does not teach the section this session belongs to
        '404':
          description: Session not found
  /teacher/sections/{sectionId}/dashboard:
    get:
      summary: Get per-student attendance averages for a section
      security:
        - BearerAuth: []
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of student attendance summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    studentId:
                      type: string
                      format: uuid
                    name:
                      type: string
                    email:
                      type: string
                    presentCount:
                      type: integer
                    totalSessions:
                      type: integer
                    attendancePct:
                      type: number
        '403':
          description: This teacher does not teach this section
        '404':
          description: Section not found
  /teacher/sections/{sectionId}/sessions:
    get:
      summary: Get past class session history for a section
      security:
        - BearerAuth: []
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of past sessions and their attendance counts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    sessionId:
                      type: string
                      format: uuid
                    startedAt:
                      type: string
                      format: date-time
                    endedAt:
                      type: string
                      format: date-time
                      nullable: true
                    presentCount:
                      type: integer
                    absentCount:
                      type: integer
        '403':
          description: This teacher does not teach this section
        '404':
          description: Section not found
...
  /student/courses:
    get:
      summary: Get student enrolled courses and current attendance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of student courses
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    sectionId:
                      type: string
                      format: uuid
                    courseName:
                      type: string
                      example: "Introduction to Computer Science"
                    courseCode:
                      type: string
                      example: "CS101"
                    presentCount:
                      type: integer
                      example: 18
                    totalSessions:
                      type: integer
                      example: 20
                    attendancePct:
                      type: number
                      example: 90.00
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    PlatformAdminKey:
      type: apiKey
      in: header
      name: X-Platform-Admin-Key